+++ src/cubeObj.h
@@ -50,7 +50,7 @@ class CubeObj {
     void changeX(int);
     void changeY(int);
     void changeZ(int);
-    void changePosTowards(CubeObj*,int);
+    void changePosTowards(CubeObj*,double);
 
     void setMomentumX(int);
     void setMomentumY(int);
